% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nel3.R
\name{buildEFOOntSupport}
\alias{buildEFOOntSupport}
\title{read and model the Experimental Factor Ontology as shipped in OWL and parsed in redland}
\usage{
buildEFOOntSupport()
}
\description{
read and model the Experimental Factor Ontology as shipped in OWL and parsed in redland
}
