% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellTypes}
\alias{CellTypes}
\title{cellTypes: data.frame with ids and terms}
\format{
TermSet instance
}
\source{
efo.owl, August 2017, subclasses of \url{http://www.ebi.ac.uk/efo/EFO_0000324}
}
\usage{
CellTypes
}
\description{
cellTypes: data.frame with ids and terms
}
\examples{
data(CellTypes)
head(slot(CellTypes, "cleanFrame"))
}
\keyword{datasets}
