% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owlSupp.R
\name{label_URL}
\alias{label_URL}
\title{acquire the label of an ontology subject tag}
\usage{
label_URL(urlstring = "<http://www.ebi.ac.uk/efo/EFO_0000311>", model, world)
}
\description{
acquire the label of an ontology subject tag
}
