% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se1.3M.R
\name{se1.3M}
\alias{se1.3M}
\title{add/retrieve HSDS-based SE to/from cache}
\usage{
se1.3M(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{BiocFileCache-like cache}
}
\description{
add/retrieve HSDS-based SE to/from cache
}
