% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nel3.R
\name{secLevGen}
\alias{secLevGen}
\title{attempt construction of a TermSet with children of a vector of terms}
\usage{
secLevGen(TermSet, choices, rrdfsupp)
}
\arguments{
\item{TermSet}{instance of \code{\link[ontoProc]{TermSet-class}}}

\item{choices}{vector of strings that should be present in cleanFrame component of \code{TermSet}}

\item{rrdfsupp}{instance of \code{\link[ontoProc]{RrdfSupport-class}}}
}
\value{
TermSet instance
}
\description{
attempt construction of a TermSet with children of a vector of terms
}
\examples{
require("ontoProc")
.efosupp = buildEFOOntSupport()
cc = children_URL("<http://www.ebi.ac.uk/efo/EFO_0000324>", .efosupp@model, .efosupp@world)
cc
secLevGen(cc, "B cell", .efosupp)
}
