% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owlSupp.R
\name{siblings_URL}
\alias{siblings_URL}
\title{generate a TermSet with siblings of a given term}
\usage{
siblings_URL(urlstring = "<http://www.ebi.ac.uk/efo/EFO_1001209>", model,
  world)
}
\arguments{
\item{urlstring}{a URI that serves as an rdf-schema class}

\item{model}{RDF model instance as defined in redland package}

\item{world}{RDF world instance as defined in redland package}
}
\description{
generate a TermSet with siblings of a given term
}
\examples{
if (!exists(".efosupp")) .efosupp = buildEFOntSupport()
siblings_URL( model=.efosupp@model, world=.efosupp@world )
}
