% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenx500}
\alias{tenx500}
\title{tenx500: serialized full SummarizedExperiment for demonstration}
\format{
SummarizedExperiment instance
}
\source{
restfulSE se1.3M pared down to 500 samples, assay materialized and assigned
}
\usage{
tenx500
}
\description{
tenx500: serialized full SummarizedExperiment for demonstration
}
\examples{
data(tenx500)
tenx500
}
\keyword{datasets}
