% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nel3.R
\name{tenXplore}
\alias{tenXplore}
\title{basic shiny interface to 10x data with ontological setup for cell selection}
\usage{
tenXplore(remouse)
}
\arguments{
\item{remouse}{optional SummarizedExperiment instance, assay data in memory}
}
\value{
shiny app invocation
}
\description{
basic shiny interface to 10x data with ontological setup for cell selection
}
\examples{
tenXplore
}
