% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenXplore.R
\name{tenXplore}
\alias{tenXplore}
\title{basic shiny interface to 10x data with ontological setup for cell selection}
\usage{
tenXplore()
}
\value{
shiny app invocation
}
\description{
basic shiny interface to 10x data with ontological setup for cell selection
}
\note{
Starts slowly as it sets up connection to HDF Server.
}
\examples{
tenXplore
}
